FIX gravity and windspeed by delta
This was a big mistake! Instead of multiplying by the expected render
time, my answer only added a pure constant. For the screen I was on this
didn't make a difference since my overpowered device can haddle the
annimation but if the fps had stumbled it would have caused weird
physics.